refacor: 각 입력 종류별 컴포넌트 분리
