+++ src/components/common/FormSubmitButtons.tsx
@@ -2,13 +2,14 @@ import Button from './Button';
 
 interface FormSubmitButtonsProps {
   onCancel: () => void;
-  onSubmit: () => void;
-  onPatch: () => void;
+  onSubmit?: () => void;  // optional로 변경
+  onPatch?: () => void;   // optional로 변경
   isEdit?: boolean;
   isDisabled: boolean;
   isLoading: boolean;
   submitLabel?: string;
   cancelLabel?: string;
+  submitType?: 'button' | 'submit';  // 추가
 }
 /**
  * 폼 제출 버튼 컴포넌트 - 리뷰, 축제 관리자 승급 신청 페이지에서 사용
@@ -30,16 +31,18 @@ const FormSubmitButtons = ({
   isLoading,
   submitLabel = isEdit ? '수정하기' : '제출하기',
   cancelLabel = '취소',
+  submitType = 'button',  // 기본값 'button'
 }: FormSubmitButtonsProps) => {
   return (
-    <div className="flex gap-3 mt-4">
+    <div className="flex gap-3 mt-4 pb-20">  {/* pb-20 추가 */}
       <Button variant="secondary" className="flex-1" onClick={onCancel}>
         {cancelLabel}
       </Button>
       <Button
+        type={submitType}  // type 동적으로 설정
         variant="primary"
         className="flex-1"
-        onClick={isEdit ? onPatch : onSubmit}
+        onClick={submitType === 'button' ? (isEdit ? onPatch : onSubmit) : undefined}
         disabled={isDisabled}
       >
         {isLoading ? (isEdit ? '수정 중...' : '제출 중...') : submitLabel}
